+++ WebApi/InWords.WebApi/InWords.WebApi/Services/UserWordPairService/Models/LicenseProviders/UncertainKnowledge.cs
@@ -10,7 +10,7 @@ namespace InWords.WebApi.Services.UserWordPairService.Models.LicenseProviders
     {
         public override KnowledgeLicense Grant(KnowledgeLicense knowledgeLicense)
         {
-            knowledgeLicense.Period = Convert.ToInt32(Math.Truncate(0.2 * knowledgeLicense.Period));
+            knowledgeLicense.Period = Convert.ToInt16(Math.Truncate(0.2 * knowledgeLicense.Period));
             return base.Grant(knowledgeLicense);
         }
     }
